etl/3_DM/src/requirements.txt
pip~=24.0
python-dotenv~=1.0.1
pymongo~=4.10.1
psycopg2>=2.9,<3.0
